requirements.txt
# Core Dependencies
torch>=2.1.0
transformers>=4.42.0
peft>=0.11.0
accelerate>=0.30.0
safetensors>=0.4.2
bitsandbytes>=0.43.0

# RAG Components
langchain>=0.1.0
langchain-community>=0.0.20
sentence-transformers>=2.3.0
faiss-cpu>=1.7.4
pypdf2>=3.0.0

# Data Processing
arxiv>=2.1.0
datasets>=2.16.0
numpy>=1.24.0
pandas>=2.0.0

# Evaluation
rouge-score>=0.1.2
nltk>=3.8.0

# UI
streamlit>=1.31.0
plotly>=5.18.0

# Utilities
python-dotenv>=1.0.0
tqdm>=4.66.0
requests>=2.31.0
